+++ rusk-wallet/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change withdraw rewards to withdraw partial amount [#2538]
+- Change transaction history fee display to be negative
+- Change transaction history item type to display public/shielded
 
 ### Fixed